+++ noctavox/src/tui/layout.rs
@@ -12,11 +12,11 @@ pub struct AppLayout {
 impl AppLayout {
     pub fn new(area: Rect, state: &mut UiState) -> Self {
         let prog_height = match state.display_progress() {
-            true => match (state.get_progress_display(), area.height > 20) {
-                (ProgressDisplay::Waveform | ProgressDisplay::Oscilloscope, true) => 6,
-                _ => 3,
-            },
             false => 0,
+            true => match (state.get_progress_display(), area.height > 20) {
+                (ProgressDisplay::ProgressBar, _) | (_, false) => 3,
+                _ => 6,
+            },
         };
 
         let search_height = match state.get_mode() == Mode::Search {